[OAP-COMMON] rename common package (#1218)
* rename common package

* change common version to oap version
