+++ oap-cache/oap/pom.xml
@@ -122,7 +122,7 @@
     <dependency>
       <groupId>com.intel</groupId>
       <artifactId>oap-common</artifactId>
-      <version>${oap.common.version}</version>
+      <version>${project.version}</version>
     </dependency>
     <dependency>
       <groupId>org.eclipse.jetty</groupId>
