+++ oap-cache/oap/src/main/scala/org/apache/spark/sql/execution/datasources/oap/filecache/MemoryManager.scala
@@ -22,7 +22,7 @@ import java.util.concurrent.atomic.AtomicLong
 
 import scala.util.Success
 
-import com.intel.oap.unsafe.PersistentMemoryPlatform
+import com.intel.oap.common.unsafe.PersistentMemoryPlatform
 
 import org.apache.spark.SparkEnv
 import org.apache.spark.internal.Logging
