+++ oap-common/pom.xml
@@ -23,7 +23,6 @@
     </parent>
 
     <artifactId>oap-common</artifactId>
-    <version>0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <dependencies>
