+++ pom.xml
@@ -24,7 +24,6 @@
         <exec.maven.version>1.6.0</exec.maven.version>
         <guava.version>18.0</guava.version>
         <java.version>1.8</java.version>
-        <oap.common.version>0.1-SNAPSHOT</oap.common.version>
         <slf4j.version>1.7.16</slf4j.version>
     </properties>
